+++ requirements.txt
@@ -6,3 +6,4 @@ PyJWT~=2.10.1
 python-dateutil~=2.9.0.post0
 pytest~=8.4.1
 python-dotenv~=1.1.1
+passlib~=1.7.4
